@@ -20,28 +20,34 @@
         "docs": "https://github.com/getgrav/grav-plugin-admin/blob/master/README.md"
     },
     "require": {
-        "php": ">=7.1.3",
+        "php": "^7.3.6 || ^8.0",
         "ext-json": "*",
-        "composer/semver": "^1.4",
-        "p3k/picofeed": "@stable"
+        "p3k/picofeed": "@stable",
+        "scssphp/scssphp": "^1.5"
     },
     "require-dev": {
         "codeception/codeception": "^2.4",
         "fzaninotto/faker": "^1.8",
-        "symfony/yaml": "~4.1",
-        "symfony/console": "~4.1",
-        "symfony/finder": "~4.1",
-        "symfony/event-dispatcher": "~4.1"
+        "symfony/yaml": "~4.4",
+        "symfony/console": "~4.4",
+        "symfony/finder": "~4.4",
+        "symfony/event-dispatcher": "~4.4"
+    },
+    "replace": {
+        "symfony/polyfill-php72": "*",
+        "symfony/polyfill-php73": "*"
     },
     "autoload": {
+        "psr-4": {
+            "Grav\\Plugin\\Admin\\": "classes/plugin"
+        },
         "classmap":  [
-            "classes/",
             "admin.php"
         ]
     },
     "config": {
         "platform": {
-            "php": "7.1.3"
+            "php": "7.3.6"
         }
     },
     "scripts": {
