@@ -16,18 +16,17 @@
     ],
     "require": {
         "php": ">=7.0.0",
-        "egulias/email-validator": "~2.0",
+        "egulias/email-validator": "^2.0",
         "symfony/polyfill-iconv": "^1.0",
         "symfony/polyfill-mbstring": "^1.0",
         "symfony/polyfill-intl-idn": "^1.10"
     },
     "require-dev": {
-        "mockery/mockery": "~0.9.1",
-        "symfony/phpunit-bridge": "^3.4.19|^4.1.8"
+        "mockery/mockery": "^1.0",
+        "symfony/phpunit-bridge": "^4.4|^5.0"
     },
     "suggest": {
-        "ext-intl": "Needed to support internationalized email addresses",
-        "true/punycode": "Needed to support internationalized email addresses, if ext-intl is not installed"
+        "ext-intl": "Needed to support internationalized email addresses"
     },
     "autoload": {
         "files": ["lib/swift_required.php"]
