@@ -20,9 +20,8 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=7.1",
         "symfony/polyfill-intl-normalizer": "^1.10",
-        "symfony/polyfill-php70": "^1.10",
         "symfony/polyfill-php72": "^1.10"
     },
     "autoload": {
@@ -35,7 +34,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "1.18-dev"
+            "dev-main": "1.22-dev"
         },
         "thanks": {
             "name": "symfony/polyfill",
