@@ -20,7 +20,7 @@
     "docs": "https://github.com/getgrav/grav-plugin-form/blob/master/README.md"
   },
   "require": {
-    "php": ">=7.1.3",
+    "php": "^7.3.6 || ^8.0",
     "ext-json": "*",
     "google/recaptcha": "^1.2"
   },
@@ -34,7 +34,7 @@
   },
   "config": {
     "platform": {
-      "php": "7.1.3"
+      "php": "7.3.6"
     }
   }
 }
