@@ -23,11 +23,10 @@
     "require": {
         "paragonie/random_compat": ">= 2",
         "ext-openssl": "*",
-        "php":  ">=5.4.0"
+        "php":  ">=5.6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4|^5",
-        "nikic/php-parser": "^2.0|^3.0|^4.0"
+        "phpunit/phpunit": "^4|^5|^6|^7|^8|^9"
     },
     "bin": [
         "bin/generate-defuse-key"
