@@ -87,7 +87,13 @@ export default class ArrayField {
 
             container.data('array-sort', new Sortable(container.get(0), {
                 handle: '.fa-bars',
-                animation: 150
+                animation: 150,
+                onUpdate: () => {
+                    const item = container.find('[data-grav-array-type="row"]:first');
+                    this._setTemplate(item);
+                    const template = item.data('array-template');
+                    this.refreshNames(template);
+                }
             }));
         });
     }
@@ -105,7 +111,13 @@ export default class ArrayField {
         let escaped_name = !template.isValueOnly() ? keyElement.val() : this.getIndexFor(element);
         escaped_name = escaped_name.toString().replace(/\[/g, '%5B').replace(/]/g, '%5D');
         let name = `${template.getName()}[${escaped_name}]`;
-        valueElement.attr('name', !valueElement.val() ? template.getName() : name);
+
+        if (!template.isValueOnly() && (!keyElement.val() && !valueElement.val())) {
+            valueElement.attr('name', '');
+        } else {
+            // valueElement.attr('name', !valueElement.val() ? template.getName() : name);
+            valueElement.attr('name', name);
+        }
 
         this.refreshNames(template);
     }
@@ -154,8 +166,8 @@ export default class ArrayField {
 
         inputs.each((index, input) => {
             input = $(input);
-            let name = input.attr('name');
-            name = name.replace(/\[\d+\]$/, `[${index}]`);
+            const preserved_name = input.closest('[data-grav-array-name]');
+            const name = `${preserved_name.attr('data-grav-array-name')}[${index}]`;
             input.attr('name', name);
         });
 
