@@ -173,7 +173,7 @@ export default class CollectionsField {
 
             item.attr('data-collection-key', hasCustomKey ? observedValue : index);
 
-            ['name', 'data-grav-field-name', 'for', 'id', 'data-grav-file-settings', 'data-grav-array-name'].forEach((prop) => {
+            ['name', 'data-grav-field-name', 'for', 'id', 'data-grav-file-settings', 'data-file-post-add', 'data-file-post-remove', 'data-grav-array-name'].forEach((prop) => {
                 item.find('[' + prop + '], [_' + prop + ']').each(function() {
                     let element = $(this);
                     let indexes = [];
@@ -186,7 +186,7 @@ export default class CollectionsField {
                     // special case to preserve array field index keys
                     if (prop === 'name' && element.data('gravArrayType')) {
                         const match_index = element.attr(prop).match(/\[[0-9]{1,}\]$/);
-                        const pattern = element.closest('[data-grav-array-name]').data('gravArrayName');
+                        const pattern = element[0].closest('[data-grav-array-name]').dataset.gravArrayName;
                         if (match_index && pattern) {
                             array_index = match_index[0];
                             element.attr(prop, `${pattern}${match_index[0]}`);
