@@ -6,12 +6,19 @@ import DateTimeField, { Instance as DateTimeFieldInstance } from './datetime';
 import EditorField, { Instance as EditorFieldInstance } from './editor';
 import ColorpickerField, { Instance as ColorpickerFieldInstance } from './colorpicker';
 import FilesField, { Instance as FilesFieldInstance } from './files';
-import MediapickerField, { Instance as MediapickerInstance } from './mediapicker';
-import MultilevelField, { Instance as MultilevelInstance } from './multilevel';
+import FolderFieldInstance from './folder';
 import SelectUniqueField, { Instance as SelectUniqueInstance } from './selectunique';
 import IconpickerField, { Instance as IconpickerInstance } from './iconpicker';
 import CronField, { Instance as CronFieldInstance } from './cron';
-import TextField, { Instance as TextFieldInstance } from './text';
+import ParentsField, { Instances as ParentsFieldInstance } from './parents';
+
+import './acl-picker';
+import './permissions';
+import './indeterminate';
+import './mediapicker';
+import './multilevel';
+import './text';
+import './range';
 
 export default {
     FilepickerField: {
@@ -46,27 +53,23 @@ export default {
         FilesField,
         Instance: FilesFieldInstance
     },
+    FolderField: {
+        Regenerate: FolderFieldInstance
+    },
     SelectUniqueField: {
         SelectUniqueField,
         Instance: SelectUniqueInstance
-    },
-    MediapickerField: {
-        MediapickerField,
-        Instance: MediapickerInstance
-    },
-    MultilevelField: {
-        MultilevelField,
-        Instance: MultilevelInstance
     },
     IconpickerField: {
         IconpickerField,
         Instance: IconpickerInstance
     },
-    CronField: {CronField,
-        Insance: CronFieldInstance
+    CronField: {
+        CronField,
+        Instance: CronFieldInstance
     },
-    TextField: {TextField,
-        Insance: TextFieldInstance
+    ParentsField: {
+        ParentsField,
+        Instance: ParentsFieldInstance
     }
 };
-
