@@ -1,14 +1,17 @@
 import $ from 'jquery';
+import { translations } from 'grav-config';
+import formatBytes from '../utils/formatbytes';
 import request from '../utils/request';
+import { Instance as Update } from './index';
 
 // Dashboard update and Grav update
-$('body').on('click', '[data-maintenance-update]', function() {
-    let element = $(this);
-    let url = element.data('maintenanceUpdate');
+$(document).on('click.remodal', '[data-remodal-id="update-grav"] [data-remodal-action="confirm"]', () => {
+    const element = $('#grav-update-button');
+    element.html(`${translations.PLUGIN_ADMIN.UPDATING_PLEASE_WAIT} ${formatBytes(Update.payload.grav.assets['grav-update'].size)}..`);
 
     element.attr('disabled', 'disabled').find('> .fa').removeClass('fa-cloud-download').addClass('fa-refresh fa-spin');
 
-    request(url, (response) => {
+    request(Update.updateURL, (response) => {
         if (response.type === 'updategrav') {
             $('[data-gpm-grav]').remove();
             $('#footer .grav-version').html(response.version);
