@@ -3,6 +3,8 @@
 namespace Grav\Plugin;
 
 use Composer\Autoload\ClassLoader;
+use DateTime;
+use Exception;
 use Grav\Common\Data\ValidationException;
 use Grav\Common\Debugger;
 use Grav\Common\Filesystem\Folder;
@@ -21,10 +23,20 @@ use Grav\Plugin\Form\Form;
 use Grav\Plugin\Form\Forms;
 use ReCaptcha\ReCaptcha;
 use ReCaptcha\RequestMethod\CurlPost;
+use RecursiveArrayIterator;
+use RecursiveIteratorIterator;
 use RocketTheme\Toolbox\File\JsonFile;
 use RocketTheme\Toolbox\File\YamlFile;
 use RocketTheme\Toolbox\File\File;
 use RocketTheme\Toolbox\Event\Event;
+use RuntimeException;
+use Twig\Extension\CoreExtension;
+use Twig\TwigFunction;
+use function count;
+use function function_exists;
+use function is_array;
+use function is_string;
+use function sprintf;
 
 /**
  * Class FormPlugin
@@ -39,19 +51,14 @@ class FormPlugin extends Plugin
 
     /** @var Form */
     protected $form;
-
     /** @var array */
     protected $forms = [];
-
     /** @var array */
     protected $flat_forms = [];
-
     /** @var array */
     protected $active_forms = [];
-
     /** @var array */
     protected $json_response = [];
-
     /** @var bool */
     protected $recache_forms = false;
 
@@ -93,8 +100,10 @@ class FormPlugin extends Plugin
 
     /**
      * Initialize forms from cache if possible
+     *
+     * @return void
      */
-    public function onPluginsInitialized()
+    public function onPluginsInitialized(): void
     {
         // Backwards compatibility for plugins that use forms.
         class_alias(Form::class, 'Grav\Plugin\Form');
@@ -134,7 +143,11 @@ class FormPlugin extends Plugin
         ]);
     }
 
-    public function onGetPageTemplates(Event $event)
+    /**
+     * @param Event $event
+     * @return void
+     */
+    public function onGetPageTemplates(Event $event): void
     {
         /** @var Types $types */
         $types = $event->types;
@@ -145,8 +158,9 @@ class FormPlugin extends Plugin
      * Process forms after page header processing, but before caching
      *
      * @param Event $e
+     * @return void
      */
-    public function onPageProcessed(Event $e)
+    public function onPageProcessed(Event $e): void
     {
         /** @var PageInterface $page */
         $page = $e['page'];
@@ -188,16 +202,20 @@ class FormPlugin extends Plugin
 
     /**
      * Initialize all the forms
+     *
+     * @return void
      */
-    public function onPagesInitialized()
+    public function onPagesInitialized(): void
     {
         $this->loadCachedForms();
     }
 
     /**
      * Catches form processing if user posts the form.
+     *
+     * @return void
      */
-    public function onPageInitialized()
+    public function onPageInitialized(): void
     {
         $submitted = false;
         $this->json_response = [];
@@ -278,7 +296,7 @@ class FormPlugin extends Plugin
                 // Discard any previously uploaded files session.
                 // and if there were any uploaded file, remove them from the filesystem
                 if ($flash = $this->grav['session']->getFlashObject('files-upload')) {
-                    $flash = new \RecursiveIteratorIterator(new \RecursiveArrayIterator($flash));
+                    $flash = new RecursiveIteratorIterator(new RecursiveArrayIterator($flash));
                     foreach ($flash as $key => $value) {
                         if ($key !== 'tmp_name') {
                             continue;
@@ -326,14 +344,16 @@ class FormPlugin extends Plugin
 
     /**
      * Add simple `forms()` Twig function
+     *
+     * @return void
      */
-    public function onTwigInitialized()
+    public function onTwigInitialized(): void
     {
         $this->grav['twig']->twig()->addFunction(
-            new \Twig_SimpleFunction('forms', [$this, 'getForm'])
+            new TwigFunction('forms', [$this, 'getForm'])
         );
 
-        $this->grav['twig']->twig()->getExtension('Twig_Extension_Core')->setEscaper('yaml', function ($twig, $string, $charset) {
+        $this->grav['twig']->twig()->getExtension(CoreExtension::class)->setEscaper('yaml', function ($twig, $string, $charset) {
             return Yaml::dump($string);
         }
         );
@@ -342,8 +362,10 @@ class FormPlugin extends Plugin
 
     /**
      * Add current directory to twig lookup paths.
+     *
+     * @return void
      */
-    public function onTwigTemplatePaths()
+    public function onTwigTemplatePaths(): void
     {
         $this->grav['twig']->twig_paths[] = __DIR__ . '/templates';
     }
@@ -352,8 +374,9 @@ class FormPlugin extends Plugin
      * Make form accessible from twig.
      *
      * @param Event $event
+     * @return void
      */
-    public function onTwigVariables(Event $event = null)
+    public function onTwigVariables(Event $event = null): void
     {
         if ($event && isset($event['page'])) {
             $page = $event['page'];
@@ -378,9 +401,10 @@ class FormPlugin extends Plugin
      * Handle form processing instructions.
      *
      * @param Event $event
-     * @throws \Exception
+     * @return void
+     * @throws Exception
      */
-    public function onFormProcessed(Event $event)
+    public function onFormProcessed(Event $event): void
     {
         /** @var Form $form */
         $form = $event['form'];
@@ -456,7 +480,7 @@ class FormPlugin extends Plugin
                 $format = $params['format'] ?? 'Y-m-d H:i:s';
                 $blueprint = $form->value()->blueprints();
                 $blueprint->set('form/fields/timestamp', ['name' => 'timestamp', 'label' => $label, 'type' => 'hidden']);
-                $now = new \DateTime('now');
+                $now = new DateTime('now');
                 $date_string = $now->format($format);
                 $form->setFields($blueprint->fields());
                 $form->setData('timestamp', $date_string);
@@ -522,7 +546,7 @@ class FormPlugin extends Plugin
                 $page = $pages->dispatch($route, true);
 
                 if (!$page) {
-                    throw new \RuntimeException('Display page not found. Please check the page exists.', 400);
+                    throw new RuntimeException('Display page not found. Please check the page exists.', 400);
                 }
 
                 unset($this->grav['page']);
@@ -551,7 +575,7 @@ class FormPlugin extends Plugin
 
                 if (!$filename) {
                     if ($operation === 'add') {
-                        throw new \RuntimeException('Form save: \'operation: add\' is only supported with a static filename');
+                        throw new RuntimeException('Form save: \'operation: add\' is only supported with a static filename');
                     }
 
                     $filename = $prefix . $this->udate($format, $raw_format) . $postfix . $ext;
@@ -574,7 +598,7 @@ class FormPlugin extends Plugin
                 if (!empty($params['raw']) || !empty($params['template'])) {
                     // Save data as it comes from the form.
                     if ($operation === 'add') {
-                        throw new \RuntimeException('Form save: \'operation: add\' is not supported for raw files');
+                        throw new RuntimeException('Form save: \'operation: add\' is not supported for raw files');
                     }
                     switch ($ext) {
                         case '.yaml':
@@ -584,7 +608,7 @@ class FormPlugin extends Plugin
                             $file = JsonFile::instance($fullFileName);
                             break;
                         default:
-                            throw new \RuntimeException('Form save: Unsupported RAW file format, please use either yaml or json');
+                            throw new RuntimeException('Form save: Unsupported RAW file format, please use either yaml or json');
                     }
 
                     $content = $form->getData();
@@ -634,7 +658,7 @@ class FormPlugin extends Plugin
 
                         if (file_exists($fullFileName)) {
                             $data = Yaml::parse($file->content());
-                            if (\count($data) > 0) {
+                            if (count($data) > 0) {
                                 array_unshift($data, $vars);
                             } else {
                                 $data[] = $vars;
@@ -650,11 +674,11 @@ class FormPlugin extends Plugin
             case 'call':
                 $callable = $params;
 
-                if (\is_array($callable) && !method_exists($callable[0], $callable[1])) {
-                    throw new \RuntimeException('Form cannot be processed (method does not exist)');
+                if (is_array($callable) && !method_exists($callable[0], $callable[1])) {
+                    throw new RuntimeException('Form cannot be processed (method does not exist)');
                 }
-                if (\is_string($callable) && !\function_exists($callable)) {
-                    throw new \RuntimeException('Form cannot be processed (function does not exist)');
+                if (is_string($callable) && !function_exists($callable)) {
+                    throw new RuntimeException('Form cannot be processed (function does not exist)');
                 }
 
                 $callable($form);
@@ -666,8 +690,9 @@ class FormPlugin extends Plugin
      * Custom field logic can go in here
      *
      * @param Event $event
+     * @return void
      */
-    public function onFormValidationProcessed(Event $event)
+    public function onFormValidationProcessed(Event $event): void
     {
         // special check for honeypot field
         foreach ($event['form']->fields() as $field) {
@@ -681,9 +706,10 @@ class FormPlugin extends Plugin
      * Handle form validation error
      *
      * @param Event $event An event object
-     * @throws \Exception
+     * @return void
+     * @throws Exception
      */
-    public function onFormValidationError(Event $event)
+    public function onFormValidationError(Event $event): void
     {
         $form = $event['form'];
         if (isset($event['message'])) {
@@ -716,6 +742,7 @@ class FormPlugin extends Plugin
      *
      * @param string|null $page_route
      * @param FormInterface|null $form
+     * @return void
      */
     public function addForm(?string $page_route, ?FormInterface $form)
     {
@@ -737,15 +764,14 @@ class FormPlugin extends Plugin
      * function to get a specific form
      *
      * @param null|array|string $data optional form `name`
-     *
      * @return FormInterface|null
      */
     public function getForm($data = null)
     {
-        if (\is_array($data)) {
+        if (is_array($data)) {
             $form_name = $data['name'] ?? null;
             $page_route = $data['route'] ?? null;
-        } elseif (\is_string($data)) {
+        } elseif (is_string($data)) {
             $form_name = $data;
             $page_route = null;
         } else {
@@ -765,12 +791,13 @@ class FormPlugin extends Plugin
                 $forms = $this->forms[$page_route];
                 $first_form = reset($forms) ?: null;
                 return $first_form;
-            } else {
-                //No form on this route. Try looking up in the current page first
-                /** @var Forms $forms */
-                $forms = $this->grav['forms'];
-                return $forms->createPageForm($this->grav['page']);
             }
+
+            // Try to get page by defined route first or get current if not found
+            $page = $this->grav['pages']->find($page_route) ?: $this->grav['page'];
+
+            // Try looking up in the defined page
+            return $this->grav['forms']->createPageForm($page);
         }
 
         // return the form you are looking for if available
@@ -786,7 +813,8 @@ class FormPlugin extends Plugin
     {
         return [
             'avatar' => [
-                'input@' => false
+                'input@' => false,
+                'media_field' => true
             ],
             'captcha' => [
                 'input@' => false
@@ -808,6 +836,7 @@ class FormPlugin extends Plugin
             ],
             'file' => [
                 'array' => true,
+                'media_field' => true,
                 'validate' => [
                     'type' => 'ignore'
                 ]
@@ -853,6 +882,7 @@ class FormPlugin extends Plugin
      * - fillWithCurrentDateTime
      *
      * @param Form $form
+     * @return void
      */
     protected function process($form)
     {
@@ -878,7 +908,7 @@ class FormPlugin extends Plugin
     /**
      * Retrieve a form based on the form name
      *
-     * @param $form_name
+     * @param string $form_name
      * @return mixed
      */
     protected function getFormByName($form_name)
@@ -953,6 +983,8 @@ class FormPlugin extends Plugin
 
     /**
      * Flatten the forms array into something that can be more easily searched
+     *
+     * @return void
      */
     protected function flattenForms()
     {
@@ -1031,13 +1063,15 @@ class FormPlugin extends Plugin
 
     /**
      * Load cached forms and merge with any currently found forms
+     *
+     * @return void
      */
     protected function loadCachedForms()
     {
         // Get and set the cache of forms if it exists
         try {
             [$forms] = $this->grav['cache']->fetch($this->getFormCacheId());
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             // Couldn't fetch cached forms.
             $forms = null;
 
@@ -1046,7 +1080,7 @@ class FormPlugin extends Plugin
             $debugger->addMessage(sprintf('Unserializing cached forms failed: %s', $e->getMessage()), 'error');
         }
 
-        if (!\is_array($forms)) {
+        if (!is_array($forms)) {
             return;
         }
 
@@ -1059,6 +1093,8 @@ class FormPlugin extends Plugin
 
     /**
      * Save the current state of the forms
+     *
+     * @return void
      */
     protected function saveCachedForms()
     {
@@ -1084,7 +1120,6 @@ class FormPlugin extends Plugin
      *
      * @param string $format
      * @param bool $raw
-     *
      * @return string
      */
     protected function udate($format = 'u', $raw = false)
@@ -1099,7 +1134,6 @@ class FormPlugin extends Plugin
         $timestamp = floor($utimestamp);
         $milliseconds = round(($utimestamp - $timestamp) * 1000000);
 
-        return date(preg_replace('`(?<!\\\\)u`', \sprintf('%06d', $milliseconds), $format), $timestamp);
+        return date(preg_replace('`(?<!\\\\)u`', sprintf('%06d', $milliseconds), $format), $timestamp);
     }
-
 }
