@@ -1,20 +1,35 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
 use Grav\Common\Language\LanguageCodes;
 use Grav\Common\Page\Page;
 use \Grav\Common\Plugin;
 
 class LangSwitcherPlugin extends Plugin
 {
+
     /**
      * @return array
      */
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized' => ['onPluginsInitialized', 0]
+            'onPluginsInitialized' => [
+                ['autoload', 100001],
+                ['onPluginsInitialized', 0]
+            ]
         ];
+    }
+
+    /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload()
+    {
+        return require __DIR__ . '/vendor/autoload.php';
     }
 
     /**
@@ -85,7 +100,7 @@ class LangSwitcherPlugin extends Plugin
 
         $data->current = $this->grav['language']->getLanguage();
 
-        $this->grav['twig']->twig_vars['langswitcher'] = $data;
+        $this->grav['twig']->twig_vars['langswitcher'] = $this->grav['langswitcher'] = $data;
 
         if ($this->config->get('plugins.langswitcher.built_in_css')) {
             $this->grav['assets']->add('plugin://langswitcher/css/langswitcher.css');
