@@ -8,7 +8,10 @@ class NoticeShortcode extends Shortcode
     public function init()
     {
         $this->shortcode->getHandlers()->add('notice', function(ShortcodeInterface $sc) {
-            $this->shortcode->addAssets('css', 'plugin://shortcode-core/css/shortcode-notice.css');
+            $css_enabled = $this->grav['config']->get('plugins.shortcode-core.css.notice_enabled', true);
+            if ($css_enabled) {
+                $this->shortcode->addAssets('css', 'plugin://shortcode-core/css/shortcode-notice.css');
+            }
             $type = $sc->getParameter('notice', $this->getBbCode($sc)) ?: 'info';
 
             return '<div class="sc-notice '.$type.'"><div>'.$sc->getContent().'</div></div>';
