@@ -2,15 +2,13 @@
 
 namespace Grav\Plugin;
 
-use Grav\Common\Config\Config;
-use Grav\Common\Data\Data;
 use Grav\Common\Page\Collection;
 use Grav\Common\Page\Page;
+use Grav\Common\Page\Pages;
 use Grav\Common\Page\Types;
 use Grav\Common\Plugin;
 use Grav\Common\Taxonomy;
 use Grav\Common\Uri;
-use Grav\Common\Utils;
 use RocketTheme\Toolbox\Event\Event;
 
 class SimplesearchPlugin extends Plugin
@@ -44,6 +42,8 @@ class SimplesearchPlugin extends Plugin
 
     /**
      * Add page template types. (for Admin plugin)
+     *
+     * @return void
      */
     public function onGetPageTemplates(Event $event)
     {
@@ -55,6 +55,8 @@ class SimplesearchPlugin extends Plugin
 
     /**
      * Add current directory to twig lookup paths.
+     *
+     * @return void
      */
     public function onTwigTemplatePaths()
     {
@@ -63,6 +65,8 @@ class SimplesearchPlugin extends Plugin
 
     /**
      * Enable search only if url matches to the configuration.
+     *
+     * @return void
      */
     public function onPluginsInitialized()
     {
@@ -79,6 +83,8 @@ class SimplesearchPlugin extends Plugin
 
     /**
      * Build search results.
+     *
+     * @return void
      */
     public function onPagesInitialized()
     {
@@ -117,6 +123,7 @@ class SimplesearchPlugin extends Plugin
         }
 
         // Explode query into multiple strings. Drop empty values
+        // @phpstan-ignore-next-line
         $this->query = array_filter(array_filter(explode(',', $query), 'trim'), 'strlen');
 
         /** @var Taxonomy $taxonomy_map */
@@ -140,8 +147,8 @@ class SimplesearchPlugin extends Plugin
             }
         }
 
-        if (!$should_process || !$filters || $query === false || (count($filters) == 1 && !reset($filters))) {
-            /** @var \Grav\Common\Page\Pages $pages */
+        if (!$should_process || !$filters || $query === false || (count($filters) === 1 && !reset($filters))) {
+            /** @var Pages $pages */
             $pages = $this->grav['pages'];
             $this->collection = $pages->all();
         } else {
@@ -179,12 +186,13 @@ class SimplesearchPlugin extends Plugin
 
         //Drop unpublished pages, but do not drop unroutable pages right now to be able to search modular pages which are unroutable per se
         $this->collection->published();
-        /** @var modular Pages $modularPageCollection */
+        /** @var Collection $modularPageCollection */
         $modularPageCollection = $this->collection->copy();
         //Get published modular pages
         $modularPageCollection->modular();
         foreach ($modularPageCollection as $cpage) {
-            if (!$cpage->parent()->published()) {
+            $parent = $cpage->parent();
+            if (!$parent || !$parent->published()) {
                 $modularPageCollection->remove($cpage);
             }
         }
@@ -262,6 +270,9 @@ class SimplesearchPlugin extends Plugin
     /**
      * Filter the pages, and return only the pages the user has access to.
      * Implementation based on Login Plugin authorizePage() function.
+     *
+     * @param Collection $collection
+     * @return Collection
      */
     public function checkForPermissions($collection)
     {
@@ -310,9 +321,9 @@ class SimplesearchPlugin extends Plugin
     }
 
     /**
-     * @param $query
+     * @param string $query
      * @param Page $page
-     * @param $taxonomies
+     * @param array|false $taxonomies
      * @return bool
      */
     private function notFound($query, $page, $taxonomies)
@@ -321,6 +332,7 @@ class SimplesearchPlugin extends Plugin
         $results = true;
         $search_content = $this->config->get('plugins.simplesearch.search_content');
 
+        $result = null;
         foreach ($searchable_types as $type => $enabled) {
             if ($type === 'title' && $enabled) {
                 $result = $this->matchText(strip_tags($page->title()), $query) === false;
@@ -344,18 +356,18 @@ class SimplesearchPlugin extends Plugin
                 }
                 $result = !$taxonomy_match;
             } elseif ($type === 'content' && $enabled) {
-                if ($search_content == 'raw') {
+                if ($search_content === 'raw') {
                     $content = $page->rawMarkdown();
                 } else {
                     $content = $page->content();
                 }
                 $result = $this->matchText(strip_tags($content), $query) === false;
-            } elseif ($type == 'header' && $enabled) {
+            } elseif ($type === 'header' && $enabled) {
                 $header = (array) $page->header();
                 $content = $this->getArrayValues($header);
                 $result = $this->matchText(strip_tags($content), $query) === false;
             }
-            $results = $results && $result;
+            $results = (bool)$result;
             if ($results === false) {
                 break;
             }
@@ -363,6 +375,11 @@ class SimplesearchPlugin extends Plugin
         return $results;
     }
 
+    /**
+     * @param string $haystack
+     * @param string $needle
+     * @return false|int
+     */
     private function matchText($haystack, $needle)
     {
         if ($this->config->get('plugins.simplesearch.ignore_accented_characters')) {
@@ -374,13 +391,15 @@ class SimplesearchPlugin extends Plugin
             }
             setlocale(LC_ALL, '');
             return $result;
-        } else {
-            return mb_stripos($haystack, $needle);
         }
+
+        return mb_stripos($haystack, $needle);
     }
 
     /**
      * Set needed variables to display the search results.
+     *
+     * @return void
      */
     public function onTwigSiteVariables()
     {
@@ -400,15 +419,22 @@ class SimplesearchPlugin extends Plugin
         }
     }
 
+    /**
+     * @param array $array
+     * @param array|null $ignore_keys
+     * @param int $level
+     * @return string
+     */
     protected function getArrayValues($array, $ignore_keys = null, $level = 0) {
         $output = '';
 
         if (is_null($ignore_keys)) {
-            $ignore_keys = $this->config()->header_keys_ignored;
+            $config = $this->config();
+            $ignore_keys = $config['header_keys_ignored'] ?? ['title', 'taxonomy','content', 'form', 'forms', 'media_order'];
         }
         foreach ($array as $key => $child) {
 
-            if ($level === 0 && in_array($key, $ignore_keys)) {
+            if ($level === 0 && in_array($key, $ignore_keys, true)) {
                 continue;
             }
 
